Factor out SPI class from ProcedureGDSFacadeSPI
